Move deploy to GitHub Pages
Move the deploy to GitHub Pages so that it will be easier to deploy it
with updates. Use GitHub Actions for the deploy to allow for more
flexibility in the future.

As far as I'm aware no scheduled build is needed to update
the website. It's most likely used now, as the server hosting the
website is not setup to receive GitHub webhooks to trigger deploys.

I found one dynamic website section, the events list on the index page.
This is updated using JavaScript loading the `events.json` file.

For this change to be merged the domain DNS needs to be updated to point
to GitHub Pages, like it does for the guides.railsgirls.com website.